Added interpolation, fixed overwrite problem.
These changes add the oft-requested interpolation feature: select any range of contiguous waves, choose Interpolate from the menu, and the range will be filled with waves interpolated from the first and last selected wave.  Also fixed is issue #17. The problem was that the same filename variable was used for both wave files and bank files, so overwrites were inevitable.